Use createNextDescribe for optimizePackageImports
