+++ test/development/basic/barrel-optimization/fixture/app/recursive-barrel-app/page.js
