+++ test/development/basic/barrel-optimization/fixture/app/recursive/page.js
