+++ test/development/basic/barrel-optimization/fixture/next.config.js
