+++ test/development/basic/barrel-optimization/fixture/pages/pages-route.js
