+++ test/development/basic/barrel-optimization/fixture/pages/visx.js
